Small refactoring

+++ src/main/java/alex/klimchuk/recipe/dto/IngredientDto.java
@@ -2,8 +2,6 @@ package alex.klimchuk.recipe.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 /**
  * Copyright Alex Klimchuk (c) 2022.
  */
@@ -17,7 +15,7 @@ public class IngredientDto {
     private Long id;
     private Long recipeId;
     private String description;
-    private BigDecimal amount;
+    private Double amount;
     private UnitOfMeasureDto unitOfMeasure;
 
 }
